// src/app/core/services/model/temperatureLevelObject.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TemperatureLevelCirculation } from './temperatureLevelCirculation';

export interface TemperatureLevelObject { 
    temperatureLevel?: TemperatureLevelCirculation;
    /**
     * only != null if TemperatureLevelCirculation = UfcManualConfig
     */
    pwhTemperatureCelsius?: number;
    /**
     * only != null if TemperatureLevelCirculation = UfcManualConfig
     */
    pwhcTemperatureCelsius?: number;
    isSelectable?: boolean;
}