// src/app/core/services/model/temperatureLevelCirculation.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type TemperatureLevelCirculation = 'DefaultNoUfc' | 'UfcLower' | 'UfcLowest' | 'UfcManualConfig';

export const TemperatureLevelCirculation = {
    DefaultNoUfc: 'DefaultNoUfc' as TemperatureLevelCirculation,
    UfcLower: 'UfcLower' as TemperatureLevelCirculation,
    UfcLowest: 'UfcLowest' as TemperatureLevelCirculation,
    UfcManualConfig: 'UfcManualConfig' as TemperatureLevelCirculation
};